# docs/openapi/token.yaml
openapi: 3.1.0
info:
  title: Go2Asia Token API
  version: 1.0.0
  description: API для работы с токенами, балансами и транзакциями
  contact:
    name: Go2Asia API Support
    email: api@go2asia.space

servers:
  - url: https://api.go2asia.space/v1/api/token
    description: Production
  - url: https://api-staging.go2asia.space/v1/api/token
    description: Staging

tags:
  - name: Balance
    description: Баланс пользователя
  - name: Transactions
    description: Транзакции

paths:
  /balance:
    get:
      tags: [Balance]
      summary: Получить баланс пользователя
      description: Возвращает текущий баланс Points и G2A токенов
      operationId: getBalance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /balance/add:
    post:
      tags: [Balance]
      summary: Добавить поинты
      description: Добавляет поинты к балансу пользователя
      operationId: addPoints
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPointsRequest'
      responses:
        '200':
          description: Поинты успешно добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /balance/subtract:
    post:
      tags: [Balance]
      summary: Списать поинты
      description: Списывает поинты с баланса пользователя
      operationId: subtractPoints
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubtractPointsRequest'
      responses:
        '200':
          description: Поинты успешно списаны
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /transactions:
    get:
      tags: [Transactions]
      summary: Получить историю транзакций
      description: Возвращает историю транзакций пользователя с пагинацией
      operationId: getTransactions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Limit:
      name: limit
      in: query
      description: Количество элементов на странице
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      description: Курсор для пагинации
      required: false
      schema:
        type: string

  schemas:
    Balance:
      type: object
      required: [points, g2a]
      properties:
        points:
          type: integer
          minimum: 0
          example: 1250
        g2a:
          type: number
          format: float
          minimum: 0
          example: 12.5

    AddPointsRequest:
      type: object
      required: [amount, reason]
      properties:
        amount:
          type: integer
          minimum: 1
          maximum: 100
          example: 100
        reason:
          type: string
          example: "quest_completed"
        metadata:
          type: object
          nullable: true
          additionalProperties: true

    SubtractPointsRequest:
      type: object
      required: [amount, reason]
      properties:
        amount:
          type: integer
          minimum: 1
          example: 50
        reason:
          type: string
          example: "voucher_purchase"
        metadata:
          type: object
          nullable: true
          additionalProperties: true

    Transaction:
      type: object
      required: [id, userId, type, amount, reason, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          format: uuid
        type:
          type: string
          enum: [points_add, points_subtract, g2a_add, g2a_subtract]
          example: "points_add"
        amount:
          type: number
          example: 100
        reason:
          type: string
          example: "quest_completed"
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        createdAt:
          type: string
          format: date-time

    TransactionsResponse:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request parameters"
            traceId:
              type: string
              format: uuid
              nullable: true

  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется авторизация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Доступ запрещён
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Конфликт
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Превышен лимит запросов
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'